# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "1.0.0",
    "title": "Movie information API",
    "description": "This API provides access to publicly available data on a limited number of movies published from the year 1990 from the Internet Movie Database and other sources. The API endpoints and their usage are described in detail below.",
    "author": "Dane Mckillop",
    "QUT student number": "n10196005",
    "signed-date": "01/06/2023"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Enter JWT Bearer token **_only_**"
      }
    },
    "schemas": {
...
      "TokenExpired": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": "true"
          },
          "message": {
            "type": "string",
            "example": "JWT token has expired"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/user/register": {
      "post": {
        "tags": ["Authentication"],
        "description": "Creates a new user account. A request body containing the user to be registered must be sent.",
        "requestBody": {
          "description": "An object containing the email and password of the user to be registered.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "mike@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "User created"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "Request body incomplete, both email and password are required"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "User already exists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "User already exists"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "example": "Too many requests, please try again later."
                }
              }
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "description": "Log in to an existing user account. A request body containing the user credentials must be sent. The longExpiry bool is a setting for development use only that makes both tokens expire after a year.",
        "tags": ["Authentication"],
        "requestBody": {
          "description": "The credentials of the user to log in.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "mike@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password"
                  },
                  "longExpiry": {
                    "type": "boolean",
                    "example": "false"
                  },
                  "bearerExpiresInSeconds": {
                    "type": "number",
                    "example": 600
                  },
                  "refreshExpiresInSeconds": {
                    "type": "number",
                    "example": 86400
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Log in successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "bearerToken": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string",
                          "example": "ajsonwebtoken"
                        },
                        "token_type": {
                          "type": "string",
                          "example": "Bearer"
                        },
                        "expires_in": {
                          "type": "number",
                          "example": 600
                        }
                      }
                    },
                    "refreshToken": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string",
                          "example": "ajsonwebtoken"
                        },
                        "token_type": {
                          "type": "string",
                          "example": "Refresh"
                        },
                        "expires_in": {
                          "type": "number",
                          "example": 86400
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid log in request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "Request body incomplete, both email and password are required"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Log in failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "Incorrect email or password"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "example": "Too many requests, please try again later."
                }
              }
            }
          }
        }
      }
    },
    "/user/refresh": {
      "post": {
        "description": "Obtain a new bearer token by using a refresh token",
        "tags": ["Authentication"],
        "requestBody": {
          "description": "The refresh token",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "example": "ajsonwebtoken"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token successfully refreshed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "bearerToken": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string",
                          "example": "ajsonwebtoken"
                        },
                        "token_type": {
                          "type": "string",
                          "example": "Bearer"
                        },
                        "expires_in": {
                          "type": "number",
                          "example": 600
                        }
                      }
                    },
                    "refreshToken": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string",
                          "example": "ajsonwebtoken"
                        },
                        "token_type": {
                          "type": "string",
                          "example": "Refresh"
                        },
                        "expires_in": {
                          "type": "number",
                          "example": 86400
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid refresh request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "Request body incomplete, refresh token required"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Click on 'Schema' below to see the possible error responses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenExpired"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "example": "Too many requests, please try again later."
                }
              }
            }
          }
        }
      }
    },
    "/user/logout": {
      "post": {
        "description": "Log the user out, invalidating the refresh token",
        "tags": ["Authentication"],
        "requestBody": {
          "description": "The refresh token",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "example": "ajsonwebtoken"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token successfully invalidated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "false"
                    },
                    "message": {
                      "type": "string",
                      "example": "Token successfully invalidated"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid refresh request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": "true"
                    },
                    "message": {
                      "type": "string",
                      "example": "Request body incomplete, refresh token required"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Click on 'Schema' below to see the possible error responses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenExpired"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "example": "Too many requests, please try again later."
                }
              }
            }
          }
        }
      }
    },
...
  }
}